requirements.txt
# TODO: Update with exact module version
numpy
torch==1.8.1
opencv_python
loguru
scikit-image
tqdm
torchvision==0.9.1
Pillow
thop
ninja
tabulate
tensorboard
lap
motmetrics
filterpy
h5py

# # verified versions
# onnx==1.8.1
# onnxruntime==1.8.0
# onnx-simplifier==0.3.5
